pyproject.toml
[project]
name = "anime-news-ci"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow[google]>=3.1.3",
    "google-cloud-aiplatform>=1.129.0",
    "httpx>=0.28.1",
]
